fix: apply pull request reviewers comments
