+++ src/database/entities/proposals-votes.entity.ts
@@ -6,19 +6,19 @@ export class ProposalsVotesEntity {
     @PrimaryColumn()
     id: string;
 
-    @Column({ type: 'int' })
+    @Column({ type: 'smallint' })
     proposal_id: number;
 
-    @Column({ type: 'varchar' })
+    @Column({ type: 'varchar', length: 42 })
     voter_address: string;
 
-    @Column({ type: 'varchar' })
+    @Column({ type: 'varchar', length: 49 })
     voter_operator_address: string;
 
-    @Column({ type: 'int' })
+    @Column({ type: 'smallint' })
     vote_option: VoteOption;
 
-    @Column({ type: 'varchar' })
+    @Column({ type: 'bigint' })
     vote_weight: string;
 
     constructor(data: Partial<ProposalsVotesEntity>) {
