+++ src/http/responses/proposal.response.ts
@@ -7,7 +7,6 @@ import Long from 'long';
 import { ProposalStatus } from '@lum-network/sdk-javascript/build/codec/cosmos/gov/v1beta1/gov';
 
 import { BalanceResponse } from '@app/http/responses/balance.response';
-import { Coin } from '@lum-network/sdk-javascript/build/codec/cosmos/base/v1beta1/coin';
 
 @Exclude()
 export class ResultResponse {
@@ -81,58 +80,6 @@ export class ProposalResponse {
     final_result: ResultResponse;
 
     constructor(data: Partial<ProposalResponse>) {
-        Object.assign(this, data);
-    }
-}
-
-export class ProposalVotersResponse {
-    @ApiProperty()
-    @Expose({ name: 'id' })
-    id: string;
-
-    @ApiProperty()
-    @Expose({ name: 'proposalId' })
-    proposal_id: string;
-
-    @ApiProperty()
-    @Expose({ name: 'voterAddress' })
-    voter_address: string;
-
-    @ApiProperty()
-    @Expose({ name: 'voterOperatorAddress' })
-    voter_operator_address: string;
-
-    @ApiProperty()
-    @Expose({ name: 'voteOption' })
-    vote_option: number;
-
-    @ApiProperty()
-    @Expose({ name: 'voteOption' })
-    vote_weight: string;
-
-    constructor(data: Partial<ProposalVotersResponse>) {
-        Object.assign(this, data);
-    }
-}
-
-export class ProposalDepositorsResponse {
-    @ApiProperty()
-    @Expose({ name: 'id' })
-    id: string;
-
-    @ApiProperty()
-    @Expose({ name: 'proposalId' })
-    proposal_id: string;
-
-    @ApiProperty()
-    @Expose({ name: 'depositorAddress' })
-    depositor_address: string;
-
-    @ApiProperty()
-    @Expose({ name: 'amount' })
-    amount: Coin[];
-
-    constructor(data: Partial<ProposalDepositorsResponse>) {
         Object.assign(this, data);
     }
 }
