+++ src/services/index.ts
@@ -5,5 +5,5 @@ export * from './stat.service';
 export * from './transaction.service';
 export * from './validator.service';
 export * from './validator-delegation.service';
-export * from './proposals-votes.service';
-export * from './proposals-deposits.service';
+export * from './proposal-vote.service';
+export * from './proposal-deposit.service';
