+++ src/services/lum-network.service.ts
@@ -1,8 +1,11 @@
 import { HttpService } from '@nestjs/axios';
 import { Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { ProposalStatus } from '@lum-network/sdk-javascript/build/codec/cosmos/gov/v1beta1/gov';
 
 import { LumClient } from '@lum-network/sdk-javascript';
+
+import moment from 'moment';
 
 @Injectable()
 export class LumNetworkService {
@@ -46,4 +49,51 @@ export class LumNetworkService {
             return [];
         }
     };
+
+    async getProposals() {
+        try {
+            // We want to sync all proposals and get the proposal_id
+            const resultsProposals = await this.client.queryClient.gov.proposals(
+                ProposalStatus.PROPOSAL_STATUS_UNSPECIFIED |
+                    ProposalStatus.PROPOSAL_STATUS_DEPOSIT_PERIOD |
+                    ProposalStatus.PROPOSAL_STATUS_VOTING_PERIOD |
+                    ProposalStatus.PROPOSAL_STATUS_PASSED |
+                    ProposalStatus.PROPOSAL_STATUS_REJECTED |
+                    ProposalStatus.PROPOSAL_STATUS_FAILED |
+                    ProposalStatus.UNRECOGNIZED,
+                '',
+                '',
+            );
+
+            return resultsProposals;
+        } catch (error) {
+            this._logger.error(`Failed to sync proposals from chain...`, error);
+        }
+    }
+
+    async getOpenVotingProposals() {
+        try {
+            const getProposals = await this.getProposals();
+
+            const now = moment();
+
+            const votingDateTime = getProposals.proposals
+                .map((el) => ({
+                    votingTime: el.votingEndTime,
+                    proposalId: el.proposalId,
+                }))
+                .filter((el) => moment(el.votingTime) > now);
+
+            if (votingDateTime.length) {
+                const getVotersByOpenProposalId = votingDateTime?.map((proposal) => proposal.proposalId).map((longInt) => longInt.low);
+                this._logger.log(`Fetched proposalId from open votes`, getVotersByOpenProposalId);
+
+                return getVotersByOpenProposalId;
+            } else {
+                this._logger.log(`No current open proposals to vote on...`);
+            }
+        } catch (error) {
+            this._logger.error(`Failed to sync proposalsById...`, error);
+        }
+    }
 }
